# contracts/openapi/data-pipeline.yaml
openapi: 3.1.0
info: {title: data-pipeline, version: 0.1.0}
paths:
  /v1/data/prepare:
    post:
      summary: 提交数据准备作业(异步;立即返回 job,不阻塞)
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PrepareJobRequest'}}}}
      responses:
        '202': {description: accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '401': {description: unauthenticated}
        '403': {description: forbidden}
  /v1/data/jobs:
    get:
      summary: 列本企业/组作业(必经 can() 按企业/组过滤;分页)
      # 实现登记:扫 JOBS_DIR 目录 O(n);作业量上千需索引(vN+,见 jobs.py list_jobs 注释)
      parameters:
        - {name: status, in: query, required: false, schema: {type: string, enum: [queued, running, succeeded, failed]}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 50, maximum: 200}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200': {description: jobs, content: {application/json: {schema: {$ref: '#/components/schemas/JobList'}}}}
        '401': {description: unauthenticated}
  /v1/data/jobs/{job_id}:
    get:
      parameters: [{name: job_id, in: path, required: true, schema: {type: string}}]
      responses:
        '200': {description: job, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '403': {description: forbidden}
        '404': {description: not found}
  /v1/data/raw:
    post:
      summary: 请求上传原始数据(can() 过 → 服务端拼 key → 返回 presigned 直传 URL;ADR-020)
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RawUploadRequest'}}}}
      responses:
        '200': {description: grant, content: {application/json: {schema: {$ref: '#/components/schemas/UploadGrant'}}}}
        '400': {description: invalid dataset/filename}
        '401': {description: unauthenticated}
        '403': {description: forbidden}
    get:
      summary: 列本企业/组原始数据集(必经 can() 过滤;分页)
      parameters:
        - {name: status, in: query, required: false, schema: {type: string, enum: [pending, ready, failed]}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 50, maximum: 200}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200': {description: list, content: {application/json: {schema: {$ref: '#/components/schemas/RawDatasetList'}}}}
        '401': {description: unauthenticated}
  /v1/data/raw/{raw_id}/complete:
    post:
      summary: 完成上传(入参仅 path id;服务端按记录再 can() + 校验对象;ADR-020 C-2)
      parameters: [{name: raw_id, in: path, required: true, schema: {type: string}}]
      requestBody: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/CompleteUploadRequest'}}}}
      responses:
        '200': {description: rawdataset, content: {application/json: {schema: {$ref: '#/components/schemas/RawDataset'}}}}
        '403': {description: forbidden}
        '404': {description: not found}
        '409': {description: object missing / complete failed}
components:
  schemas:
    PrepareJobRequest:
      type: object
      required: [dataset, source_dataset]                                          # owner 模型(ADR-024):去 group_id
      properties:
        dataset: {type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,63}$'}            # 产出(processed)数据集名
        source_dataset: {type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,63}$'}     # S2a catalog-driven:源 raw 数据集名;服务端 submit 时(带 bearer)查 metadata 解析 location(kind 须 raw)
        np: {type: [integer, 'null'], minimum: 1}
        process: {type: [array, 'null'], items: {type: object}}   # Layer 1 DJ 算子(spec §8;build_recipe(process=…))
    Job:
      type: object
      required: [id, status, terminal, dataset, enterprise_id]                     # owner 模型(ADR-024):去 group_id required,加 owner_user
      properties:
        id: {type: string}
        status: {type: string, enum: [queued, running, succeeded, failed]}
        terminal: {type: boolean}     # 派生:status∈{succeeded,failed}。客户端按此判终态(ADR-018 不变量 4;S2a 加新终态如 cancelled 不破坏轮询)
        dataset: {type: string}
        enterprise_id: {type: string, pattern: '^[a-z][a-z0-9-]{3,}$'}
        owner_user: {type: [string, 'null']}   # owner(ADR-024):提交作业的用户(sub,§1.4 不透明)
        source_dataset: {type: [string, 'null']}   # 血缘:真实来源 raw 数据集名(US3-AC1/SC-003);前端注册产物 derived_from 取它。非 required(旧 job 无则 null)
        rows_in: {type: [integer, 'null']}
        rows_written: {type: [integer, 'null']}
        lance_uri: {type: [string, 'null']}
        error: {type: [string, 'null']}
        created_at: {type: [string, 'null']}
        updated_at: {type: [string, 'null']}
    JobList:
      type: object
      required: [jobs, total]
      properties:
        jobs: {type: array, items: {$ref: '#/components/schemas/Job'}}
        total: {type: integer}     # 过滤后(can() + status)总数,非全量
    RawUploadRequest:
      type: object
      required: [dataset, filename]                                          # owner 模型(ADR-024):去 group_id
      properties:
        dataset: {type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,63}$'}        # 数据集名(=raw 前缀段)
        filename: {type: string}   # 单段文件名;服务端 raw_object_key 为 C-1 权威校验(无 '/' 防穿越,首字符非 '.'),非法→400(ADR-020 C-1)
        multipart: {type: boolean, default: false}
        parts: {type: [integer, 'null'], minimum: 1, maximum: 10000}           # multipart 时分片数
    UploadGrant:
      type: object
      required: [raw_id, oss_key, expires_in]
      properties:
        raw_id: {type: string}
        oss_key: {type: string}
        url: {type: [string, 'null']}                                          # 单传:一条 presigned PUT URL
        upload_id: {type: [string, 'null']}                                    # multipart:OSS UploadId
        part_urls: {type: [array, 'null'], items: {type: string}}             # multipart:逐片 presigned URL(序与 PartNumber 对应)
        expires_in: {type: integer}                                            # 秒,TTL≤900(ADR-020 §4)
    CompletePart:
      type: object
      required: [part_number, etag]
      properties:
        part_number: {type: integer, minimum: 1}
        etag: {type: string}
    CompleteUploadRequest:
      type: object
      properties:
        parts: {type: [array, 'null'], items: {$ref: '#/components/schemas/CompletePart'}}   # 单传留空;multipart 带各片 ETag
    RawDataset:
      type: object
      required: [id, name, enterprise_id, oss_key, status]                  # owner 模型(ADR-024):去 group_id required,加 owner_user
      properties:
        id: {type: string}
        name: {type: string}
        enterprise_id: {type: string, pattern: '^[a-z][a-z0-9-]{3,}$'}
        owner_user: {type: [string, 'null']}   # owner(ADR-024):上传用户(sub,§1.4 不透明)
        oss_key: {type: string}
        status: {type: string, enum: [pending, ready, failed]}
        size: {type: [integer, 'null']}
        error: {type: [string, 'null']}
        created_at: {type: [string, 'null']}
        updated_at: {type: [string, 'null']}
    RawDatasetList:
      type: object
      required: [raw, total]
      properties:
        raw: {type: array, items: {$ref: '#/components/schemas/RawDataset'}}
        total: {type: integer}     # 过滤后(can()+status)总数