# libs/contracts_gen/data_pipeline_models.py
# generated by datamodel-codegen:
#   filename:  data-pipeline.yaml

from __future__ import annotations

from enum import Enum
from typing import Any

from pydantic import BaseModel, conint, constr


class PrepareJobRequest(BaseModel):
    dataset: constr(pattern=r'^[a-z0-9][a-z0-9_-]{0,63}$')
    source_dataset: constr(pattern=r'^[a-z0-9][a-z0-9_-]{0,63}$')
    np: conint(ge=1) | None = None
    process: list[dict[str, Any]] | None = None


class Status(Enum):
    queued = 'queued'
    running = 'running'
    succeeded = 'succeeded'
    failed = 'failed'


class Job(BaseModel):
    id: str
    status: Status
    terminal: bool
    dataset: str
    enterprise_id: constr(pattern=r'^[a-z][a-z0-9-]{3,}$')
    owner_user: str | None = None
    source_dataset: str | None = None
    rows_in: int | None = None
    rows_written: int | None = None
    lance_uri: str | None = None
    error: str | None = None
    created_at: str | None = None
    updated_at: str | None = None


class JobList(BaseModel):
    jobs: list[Job]
    total: int


class RawUploadRequest(BaseModel):
    dataset: constr(pattern=r'^[a-z0-9][a-z0-9_-]{0,63}$')
    filename: str
    multipart: bool | None = False
    parts: conint(ge=1, le=10000) | None = None


class UploadGrant(BaseModel):
    raw_id: str
    oss_key: str
    url: str | None = None
    upload_id: str | None = None
    part_urls: list[str] | None = None
    expires_in: int


class CompletePart(BaseModel):
    part_number: conint(ge=1)
    etag: str


class CompleteUploadRequest(BaseModel):
    parts: list[CompletePart] | None = None


class Status1(Enum):
    pending = 'pending'
    ready = 'ready'
    failed = 'failed'


class RawDataset(BaseModel):
    id: str
    name: str
    enterprise_id: constr(pattern=r'^[a-z][a-z0-9-]{3,}$')
    owner_user: str | None = None
    oss_key: str
    status: Status1
    size: int | None = None
    error: str | None = None
    created_at: str | None = None
    updated_at: str | None = None


class RawDatasetList(BaseModel):
    raw: list[RawDataset]
    total: int