# services/data_pipeline_service/app.py
from __future__ import annotations
import uuid
from datetime import datetime, timezone
from fastapi import Depends, HTTPException, Query, Request
from fastapi.responses import JSONResponse
from libs.authz.engine import can
from libs.authz.types import Resource
from libs.audit.oss_audit import AuditWriter, AuditEvent
from libs.contracts_gen.data_pipeline_models import (
    PrepareJobRequest, RawUploadRequest, CompleteUploadRequest)
from libs.identity.context import Context
from libs.identity.ids import EnterpriseId
from services._scaffold.app import make_service_app
from services._scaffold.auth import context_from_request, enterprise_of   # enterprise_of 本 plan 抽自 metadata-service
from services.data_pipeline_service.jobs import JobSpec
from services.data_pipeline_service.metadata_client import MetadataClient
from services.data_pipeline_service.upload import ObjectMissing

def _audit(audit: AuditWriter, ctx: Context, ent: str, resource_uri: str,
           action: str, decision: str, reason: str, metadata: dict | None = None) -> None:
    audit.write(AuditEvent(ts=datetime.now(timezone.utc).isoformat(), enterprise_id=ent,
                           actor_user=ctx.user, actor_role=ctx.role_in(EnterpriseId(ent)) or "none",
                           action=action, resource_uri=resource_uri, decision=decision,
                           override=False, reason=reason, metadata=metadata or {}))

def build_app(runner, audit: AuditWriter, uploader=None, metadata=None):
    app = make_service_app(title="data-pipeline-service", version="0.1.0")
    metadata = metadata or MetadataClient()   # 延迟构造(默认参数 None)避免 import 期连接

    @app.post("/v1/data/prepare", status_code=202)
    def prepare(body: PrepareJobRequest, request: Request,
                ctx: Context = Depends(context_from_request)):
        ent = enterprise_of(ctx)
        # owner 模型(ADR-024):产出归提交用户(owner=ctx.user);can() 按企业隔离 + owner 判
        res = Resource(kind="dataset", enterprise_id=EnterpriseId(ent), owner=ctx.user)
        d = can(ctx, "data.prepare", res)
        if not d.allow:                       # deny → 零副作用 + 审计
            _audit(audit, ctx, ent, f"dataset/{body.dataset}", "data.prepare", "deny", d.reason)
            return JSONResponse(status_code=403, content={"reason": d.reason})
        # catalog-driven:submit 时(带 caller bearer,经 metadata can())按名解析源 raw 数据集的 OSS location。
        # worker detached 无 bearer,故必须在此解析(探针结论)。kind 须 raw(v1 只从原始数据集产出)。
        bearer = request.headers.get("authorization", "")
        try:
            ds = metadata.get_dataset("data", "datasets", body.source_dataset, bearer=bearer)
        except Exception:
            return JSONResponse(status_code=400, content={"reason": "源数据集不存在或不可读"})
        if ds.get("kind") != "raw":
            return JSONResponse(status_code=400, content={"reason": "源须为原始数据集(v1)"})
        source_location = ds["location"]
        job_id = "job-" + uuid.uuid4().hex[:16]      # 服务端不透明 id(不变量 3)
        spec = JobSpec(job_id=job_id, dataset=body.dataset, enterprise_id=ent,
                       role=ctx.role_in(EnterpriseId(ent)) or "member",   # owner 模型:企业级角色快照(无组)
                       sub=ctx.user, source_location=source_location, source_dataset=body.source_dataset,
                       np=body.np or 3, process=body.process)
        runner.submit(spec)
        return runner.get(job_id)

    @app.get("/v1/data/jobs")
    def list_jobs(status: str | None = None,
                  limit: int = Query(50, ge=1, le=200),     # 契约 maximum=200,强制(默认不验 schema 边界)
                  offset: int = Query(0, ge=0),             # 防负 offset 取尾片(契约 default=0)
                  ctx: Context = Depends(context_from_request)):
        ent = enterprise_of(ctx)
        visible = []
        for j in runner.list_jobs():                       # 纯取数(含 enterprise_id/owner_user 投影)
            je, jo = j.get("enterprise_id"), j.get("owner_user")
            if je is None or je != ent:                    # I-2 fail-closed:spec 缺失/跨企业 → 跳过
                continue
            res = Resource(kind="job", enterprise_id=EnterpriseId(ent), owner=jo)
            if not can(ctx, "data.read", res).allow:       # owner 模型:必经 can() 按 owner 过滤(非 owner 不可见)
                continue
            if status and j.get("status") != status:       # status 过滤
                continue
            visible.append(j)
        total = len(visible)                               # 过滤后总数(非全量、非页大小)
        return {"jobs": visible[offset: offset + limit], "total": total}

    @app.get("/v1/data/jobs/{job_id}")
    def get_job(job_id: str, ctx: Context = Depends(context_from_request)):
        ent = enterprise_of(ctx)
        job = runner.get(job_id)
        if job is None or job["enterprise_id"] != ent:
            raise HTTPException(status_code=404, detail="not found")   # 跨企业=不存在(不泄漏)
        res = Resource(kind="job", enterprise_id=EnterpriseId(ent), owner=job.get("owner_user"))
        d = can(ctx, "data.read", res)
        if not d.allow:                                               # 非 owner → 403(owner 模型)
            return JSONResponse(status_code=403, content={"reason": d.reason})
        return job

    @app.post("/v1/data/raw")
    def request_upload(body: RawUploadRequest, ctx: Context = Depends(context_from_request)):
        if uploader is None:
            raise HTTPException(status_code=503, detail="upload not configured")
        ent = enterprise_of(ctx)
        # owner 模型(ADR-024):上传数据归上传用户(owner=ctx.user);can() 按企业隔离 + owner 判
        res = Resource(kind="dataset", enterprise_id=EnterpriseId(ent), owner=ctx.user)
        d = can(ctx, "data.upload", res)
        if not d.allow:                                       # deny → 零副作用 + 审计
            _audit(audit, ctx, ent, f"raw/{body.dataset}", "data.upload", "deny", d.reason)
            return JSONResponse(status_code=403, content={"reason": d.reason})
        try:
            grant = uploader.create_grant(name=body.dataset, enterprise_id=ent, user_id=ctx.user,
                                          sub=ctx.user, filename=body.filename,
                                          multipart=bool(body.multipart), parts=body.parts)
        except ValueError as e:                               # 文件名/数据集名非法
            return JSONResponse(status_code=400, content={"reason": str(e)})
        # presign allow 审计带 key+TTL+raw_id(ADR-020 I-2):GC 扫 pending 超时时可据此对账
        # "授权了但结果未知"的中间态(字节已落 OSS 但 complete 丢失)。
        _audit(audit, ctx, ent, f"raw/{body.dataset}", "data.upload", "allow", "",
               metadata={"raw_id": grant["raw_id"], "oss_key": grant["oss_key"], "expires_in": grant["expires_in"]})
        return grant

    @app.post("/v1/data/raw/{raw_id}/complete")
    def complete_upload(raw_id: str, body: CompleteUploadRequest | None = None,
                        ctx: Context = Depends(context_from_request)):
        if uploader is None:
            raise HTTPException(status_code=503, detail="upload not configured")
        ent = enterprise_of(ctx)
        rec = uploader.get_record(raw_id)
        if rec is None or rec["enterprise_id"] != ent:        # 跨企业=不存在(不泄漏)
            raise HTTPException(status_code=404, detail="not found")
        res = Resource(kind="dataset", enterprise_id=EnterpriseId(ent), owner=rec.get("owner_user"))
        d = can(ctx, "data.upload", res)
        if not d.allow:                                       # 按记录 eid/owner 再 can()(C-2;owner 模型)
            return JSONResponse(status_code=403, content={"reason": d.reason})
        parts = [p.model_dump() for p in (body.parts or [])] if body and body.parts else None
        try:
            out = uploader.finalize(raw_id, parts=parts)
        except ObjectMissing as e:
            _audit(audit, ctx, ent, f"raw/{rec['name']}", "data.upload", "deny", "object-missing")
            return JSONResponse(status_code=409, content={"reason": str(e)})
        _audit(audit, ctx, ent, f"raw/{rec['name']}", "data.upload", "allow", "complete",
               metadata={"raw_id": raw_id, "oss_key": rec["oss_key"], "size": out.get("size")})
        return out

    @app.get("/v1/data/raw")
    def list_raw(status: str | None = None,
                 limit: int = Query(50, ge=1, le=200), offset: int = Query(0, ge=0),
                 ctx: Context = Depends(context_from_request)):
        if uploader is None:
            raise HTTPException(status_code=503, detail="upload not configured")
        ent = enterprise_of(ctx)
        visible = []
        for r in uploader.list_raw():
            re_, ro = r.get("enterprise_id"), r.get("owner_user")
            if re_ is None or re_ != ent:                     # fail-closed:缺失/跨企业 → 跳过
                continue
            res = Resource(kind="dataset", enterprise_id=EnterpriseId(ent), owner=ro)
            if not can(ctx, "data.read", res).allow:          # owner 模型:必经 can() 按 owner 过滤
                continue
            if status and r.get("status") != status:
                continue
            visible.append(r)
        return {"raw": visible[offset: offset + limit], "total": len(visible)}

    return app